MainMultiPlayHandler.cs: 데이터타입을 v2에 맞춰 수정

+++ Assets/Scripts/UI/Main/MainMultiPlayHandler.cs
@@ -38,7 +38,7 @@ public class MainMultiPlayHandler : MonoBehaviour
         {
             Player = new Protocol.Player
             {
-                Id = (ulong)NetManager.Instance._playerId,
+                Id = NetManager.Instance._playerId,
                 Name = NetManager.Instance.PlayerName ?? "",
                 Tag = NetManager.Instance.PlayerTag
             },
